refactor: extract transaction cash flow chart calculations
