+++ frontend/src/transactions/components/topBand/DailyCashFlowChart.tsx
@@ -31,20 +31,21 @@ import IconTooltip from '@/components/IconTooltip'
 import { formatMissingFxPairs, getFxStatusTone } from '@/utils/fxStatus'
 import { formatCurrency } from '@/utils/formatCurrency'
 import { PLACEHOLDER_DAILY_FLOW } from '@/transactions/components/topBand/constants'
-import { parseYmdLocal } from '@/transactions/utils/date'
+import {
+  DAILY_CASH_FLOW_CHART_MARGIN,
+  DAILY_CASH_FLOW_X_AXIS_PADDING,
+  getDailyCashFlowCadenceTitle,
+  getDailyCashFlowCalculation,
+  getDailyCashFlowGranularity,
+  getDailyCashFlowSeries,
+  getDailyCashFlowXAxisTickCount,
+  getDailyCashFlowXAxisTicks,
+  type DailyCashFlowChartMode,
+  type DailyCashFlowPoint,
+} from '@/transactions/utils/dailyCashFlowChart'
 import { getCashFlowFxStatusMessage } from '@/transactions/utils/fxTooltipMessages'
 
-type DailyCashFlowPoint = {
-  key: string
-  date: string
-  rangeLabel: string
-  inflow: number
-  outflow: number
-  net: number
-}
-
-export type DailyCashFlowChartMode = 'net' | 'gross'
-type DailyCashFlowGranularity = 'day' | 'week' | 'month'
+export type { DailyCashFlowChartMode } from '@/transactions/utils/dailyCashFlowChart'
 
 const titleWordTransition = { duration: 0.34, ease: [0.16, 1, 0.3, 1] } as const
 const titleWidthTransition = { duration: 0.3, ease: [0.16, 1, 0.3, 1] } as const
@@ -59,183 +60,8 @@ const titleCharVariants = {
   enter: { y: 0, opacity: 1, filter: 'blur(0px)' },
   exit: { y: '-0.7em', opacity: 0, filter: 'blur(2px)' },
 } as const
-const dailyCashFlowRangeDayCount = 31
-const weeklyCashFlowRangeDayCount = 183
-const dailyCashFlowMaxXAxisTickCount = 10
-const dailyCashFlowXAxisTickSpacing = 64
-const dailyCashFlowChartMargin = { top: 4, right: 12, bottom: 0, left: 12 } as const
-const dailyCashFlowXAxisPadding = { left: 20, right: 20 } as const
-const dailyCashFlowXAxisCandidateSteps = [1, 2, 3, 4, 5, 7, 10, 14, 15, 21, 30] as const
-
 // Recharts runtime accepts cubic-bezier strings, but Area's public type only lists preset names
 const chartAnimationEasing = 'cubic-bezier(0.05,0.025,0.41,0.941)' as 'ease-in-out'
-
-function formatYmdLocal(date: Date) {
-  return [
-    date.getFullYear(),
-    String(date.getMonth() + 1).padStart(2, '0'),
-    String(date.getDate()).padStart(2, '0'),
-  ].join('-')
-}
-
-function getDailyCashFlowGranularity(fromDate: string, toDate: string): DailyCashFlowGranularity {
-  if (fromDate > toDate) return 'day'
-
-  const from = parseYmdLocal(fromDate)
-  const to = parseYmdLocal(toDate)
-  const dayCount = Math.max(
-    1,
-    Math.round((to.getTime() - from.getTime()) / 86400000) + 1,
-  )
-
-  if (dayCount <= dailyCashFlowRangeDayCount) return 'day'
-  if (dayCount <= weeklyCashFlowRangeDayCount) return 'week'
-  return 'month'
-}
-
-function getDailyCashFlowCadenceTitle(granularity: DailyCashFlowGranularity) {
-  if (granularity === 'week') return 'Weekly'
-  if (granularity === 'month') return 'Monthly'
-  return 'Daily'
-}
-
-function getDailyCashFlowPeriodName(granularity: DailyCashFlowGranularity) {
-  if (granularity === 'week') return 'week'
-  if (granularity === 'month') return 'month'
-  return 'day'
-}
-
-function getDailyCashFlowCalculation(
-  granularity: DailyCashFlowGranularity,
-  mode: DailyCashFlowChartMode,
-) {
-  const period = getDailyCashFlowPeriodName(granularity)
-  return mode === 'net'
-    ? `Each ${period}'s money in minus money out. Transfers count except Balance Adjustment.`
-    : `Each ${period}'s money in and money out. Transfers count except Balance Adjustment.`
-}
-
-function formatCashFlowPointLabel(date: Date, granularity: DailyCashFlowGranularity) {
-  return date.toLocaleDateString('en-US', {
-    month: 'short',
-    day: granularity === 'month' ? undefined : 'numeric',
-  })
-}
-
-function formatCashFlowTooltipDate(date: Date) {
-  return date.toLocaleDateString('en-US', {
-    year: 'numeric',
-    month: 'short',
-    day: 'numeric',
-  })
-}
-
-function formatCashFlowRangeLabel(start: Date, end: Date, granularity: DailyCashFlowGranularity) {
-  if (granularity === 'day' || formatYmdLocal(start) === formatYmdLocal(end)) {
-    return formatCashFlowTooltipDate(start)
-  }
-
-  return `${formatCashFlowTooltipDate(start)} - ${formatCashFlowTooltipDate(end)}`
-}
-
-function getDailyCashFlowSeries(
-  raw: DailyCashFlow[],
-  granularity: DailyCashFlowGranularity,
-): DailyCashFlowPoint[] {
-  return raw.map((entry) => {
-    const bucketStart = parseYmdLocal(entry.date)
-    const bucketEnd = parseYmdLocal(entry.end_date)
-
-    return {
-      key: entry.date,
-      date: formatCashFlowPointLabel(bucketStart, granularity),
-      rangeLabel: formatCashFlowRangeLabel(bucketStart, bucketEnd, granularity),
-      inflow: entry.inflow,
-      outflow: entry.outflow,
-      net: entry.inflow + entry.outflow,
-    }
-  })
-}
-
-function getDailyCashFlowXAxisTickCount(chartWidth: number | undefined) {
-  if (chartWidth === undefined) return dailyCashFlowMaxXAxisTickCount
-
-  const usableWidth = Math.max(
-    chartWidth
-      - dailyCashFlowChartMargin.left
-      - dailyCashFlowChartMargin.right
-      - dailyCashFlowXAxisPadding.left
-      - dailyCashFlowXAxisPadding.right,
-    0,
-  )
-
-  return Math.max(
-    2,
-    Math.min(
-      dailyCashFlowMaxXAxisTickCount,
-      Math.floor(usableWidth / dailyCashFlowXAxisTickSpacing) + 1,
-    ),
-  )
-}
-
-function getDailyCashFlowXAxisTickIndexesForStep(dataLength: number, step: number) {
-  const lastIndex = dataLength - 1
-  const indexes: number[] = []
-
-  for (let index = 0; index < lastIndex; index += step) {
-    indexes.push(index)
-  }
-
-  const finalGap = lastIndex - indexes[indexes.length - 1]
-  if (finalGap === 0) return indexes
-
-  if (indexes.length > 1 && finalGap < step / 2) {
-    indexes[indexes.length - 1] = lastIndex
-    return indexes
-  }
-
-  return [...indexes, lastIndex]
-}
-
-function getDailyCashFlowXAxisTickIndexes(dataLength: number, maxTickCount: number) {
-  const cappedTickCount = Math.min(maxTickCount, dataLength)
-  if (cappedTickCount === 0) return []
-  if (cappedTickCount === 1) return [0]
-
-  const lastIndex = dataLength - 1
-  const minimumStep = Math.max(1, Math.ceil(lastIndex / (cappedTickCount - 1)))
-  const candidateSteps = dailyCashFlowXAxisCandidateSteps.some((step) => step === minimumStep)
-    ? dailyCashFlowXAxisCandidateSteps
-    : [...dailyCashFlowXAxisCandidateSteps, minimumStep].sort((a, b) => a - b)
-
-  let bestIndexes = [0, lastIndex]
-  let bestScore = Number.POSITIVE_INFINITY
-
-  for (const step of candidateSteps) {
-    if (step < minimumStep) continue
-
-    const indexes = getDailyCashFlowXAxisTickIndexesForStep(dataLength, step)
-    if (indexes.length > cappedTickCount) continue
-
-    const gaps = indexes.slice(1).map((index, gapIndex) => index - indexes[gapIndex])
-    const gapSpread = Math.max(...gaps) - Math.min(...gaps)
-    const unusedTickPenalty = (cappedTickCount - indexes.length) * 0.2
-    const score = gapSpread / step + unusedTickPenalty
-
-    if (score < bestScore) {
-      bestIndexes = indexes
-      bestScore = score
-    }
-  }
-
-  return bestIndexes
-}
-
-function getDailyCashFlowXAxisTicks(data: DailyCashFlowPoint[], maxTickCount: number) {
-  return getDailyCashFlowXAxisTickIndexes(data.length, maxTickCount).map((index) => (
-    data[index].key
-  ))
-}
 
 function getDailyCashFlowTooltipKey(point: DailyCashFlowPoint) {
   return point.key
@@ -485,7 +311,7 @@ export default function DailyCashFlowChart({
           <AreaChart
             key={`daily-flow-${mode}-${chartAnimationKey}`}
             data={dailyFlow}
-            margin={dailyCashFlowChartMargin}
+            margin={DAILY_CASH_FLOW_CHART_MARGIN}
             onMouseMove={(state, event) => showDailyCashFlowTooltip(state, event)}
             onMouseLeave={hideDailyCashFlowTooltip}
           >
@@ -509,7 +335,7 @@ export default function DailyCashFlowChart({
               tickFormatter={(value) => dailyFlowLabelsByKey.get(String(value)) ?? String(value)}
               axisLine={false}
               tickLine={false}
-              padding={dailyCashFlowXAxisPadding}
+              padding={DAILY_CASH_FLOW_X_AXIS_PADDING}
               interval={0}
               ticks={dailyFlowXAxisTicks}
             />
